refactor(Edificio): change id type
Change id type from long to int.

+++ backend/src/main/java/org/ufromap/models/Edificio.java
@@ -3,14 +3,14 @@ package org.ufromap.models;
 import java.util.List;
 
 public class Edificio{
-    private final long id;
+    private final int id;
     private final String nombre;
     private final String alias;
     private final String ubicacion;
     private final String tipo;
     private final List<Sala> salas;
 
-    public Edificio(long id, String nombre, String alias, String ubicacion, String tipo, List<Sala> salas) {
+    public Edificio(int id, String nombre, String alias, String ubicacion, String tipo, List<Sala> salas) {
         this.id = id;
         this.nombre = nombre;
         this.alias = alias;
@@ -19,7 +19,7 @@ public class Edificio{
         this.salas = salas;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
